PresentableObject
Click on the object (like a letter) to expand it on the screen. Then click and drag to scroll up or down. Press space and click to place the object back.

Currently, the space is only temporary - I need to move the listener in the Clickable object class to OnMouseUp in order to implement a timing system where long presses will trigger scrolling. When scrolling, the mouse cursor will be locked to the center. There is a fix, but it will likely break the game on non-windows computers.